fix: suppress GCC -Wsign-compare and -Wmisleading-indentation warnings
XLComments.cpp: cast right-hand side of MAX_ROWS comparison to uint32_t
to eliminate signed/unsigned comparison warning.

XLStyles_CellFormats.cpp: restructure setXfId() with explicit if/else
braces so that the throw is reachable when m_permitXfId is false,
eliminating the misleading-indentation warning and correcting the logic.

+++ OpenXLSX/sources/XLComments.cpp
@@ -375,7 +375,7 @@ void XLComments::setupVmlShape(const std::string& cellRef, uint32_t destRow, uin
         }
 
         uint32_t anchorTopRow, anchorBottomRow;
-        if (OpenXLSX::MAX_ROWS - destRow > topRowOffset + heightRows) {
+        if (OpenXLSX::MAX_ROWS - destRow > static_cast<uint32_t>(topRowOffset + heightRows)) {
             anchorTopRow    = (destRow - 1) + topRowOffset;
             anchorBottomRow = (destRow - 1) + topRowOffset + heightRows;
         }
